Refactor benchmark profiles and add new sum benchmark

+++ tensor-bench/benches/binary/add_broadcast_f16.rs
@@ -3,7 +3,6 @@ use std::time::Duration;
 use criterion::{ black_box, criterion_group, criterion_main, BenchmarkId, Criterion };
 use tch::{ Tensor, Kind, Device };
 use tensor_dyn::{ tensor_base::_Tensor, Random };
-use half::f16;
 
 fn add_f16_benchmark(c: &mut Criterion) {
     tensor_dyn::set_num_threads(num_cpus::get_physical());
@@ -25,10 +24,10 @@ fn add_f16_benchmark(c: &mut Criterion) {
         let mut shape2 = shapes[idx];
         shape2[1] = 1;
         shape2[3] = 1;
-        let a = black_box(Tensor::randn(shape1, (Kind::Half, Device::Cpu)));
-        let c = black_box(Tensor::randn(shape2, (Kind::Half, Device::Cpu)));
-        let a2 = black_box(_Tensor::<f16>::randn(shape1).unwrap());
-        let c2 = black_box(_Tensor::<f16>::randn(shape2).unwrap());
+        let a = black_box(Tensor::randn(shape1, (Kind::Float, Device::Cpu)));
+        let c = black_box(Tensor::randn(shape2, (Kind::Float, Device::Cpu)));
+        let a2 = black_box(_Tensor::<f32>::randn(shape1).unwrap());
+        let c2 = black_box(_Tensor::<f32>::randn(shape2).unwrap());
         group.bench_with_input(BenchmarkId::new("torch", idx), &shapes[idx], |b, _| {
             b.iter(|| { &a + &c });
         });
